refactor: modularize script.js into separate feature-based modules
